Experience System + Basic Screen Fade Transition
Added the experience system. It uses interfaces ICanBeLeveled and IGiveExperience. As of this update, party members can gain experience from enemies, and the amount of exp per enemy is based on the ratio of the party member level to the enemy level and the enemy's base experience. Leveling up also works and calls an action that can be used to give level up rewards and for UI calls.