Global Variables Upgrade + Class System + Animations + Editor Updates
Updated Global Variables to create an instance of itself at runtime, so now any script can access a global variable at runtime without global variables needing to be referenced.  Created the class system. At this moment, you can assign a class tier, a list of abilities that can be unlocked along with the level requirement, a party member that can use the class, and each class has a different base exp that is assigned to its enum. Growth rates are implemented and fully working as well!. The functionality for stat bonuses is there but not yet implemented. Also implemented the new special attack animation for Lilith and her overworld animations + added Alek's overworld idle animation. Finally, i updated the Scriptable Object Editor to be able to create more scriptable objects and moved the scriptable objects not in that folder to the scriptable object folder.