Player Positon Tracking System
Made a (surprisingly complex) system that tracks the players positon when going into a battle. If the player is transitioning between areas, a spawn id is set and the corresponding spawn point is used to spawn the player.  A player spawn point has an id that should be the same for any 2 connecting spawn points.