Basic Level Up Results Sequence
Created sytem to show each character that leveled up and their stats at the end of battle. This is very much not perfect. There is no flair to the sequence. It does not show each stat increase in a sequence like Fire Emblem. The actual level up panel is ugly. Still haven't implemented health increases. A lot of logic is in the win sequence, which I don't like.