Value Dropdowns and (Hopefully) Bug Fix
Added value dropdowns to the rotation and position manager that gets all of the possible values from the scene, which will make testing SOO much easier. Also hopefully fixed bug where battles won't actually load. I think this was because of multiple enemies triggering it. I put the timescale to 0 when it triggers, which seems to be working.